+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
-- None
+- Changed SwiftUI's `Label` initializer to take an optional `SFSymbol` (By [Steven Sorial](https://github.com/StevenSorial))
 
 ### Fixed